requirements.txt
# 核心套件
google-generativeai==0.8.3
flask==3.0.0
line-bot-sdk==3.12.0
pillow==10.4.0

# Google Cloud AI 服務
google-cloud-aiplatform>=1.38.0
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.14.0

# 工具套件
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-storage>=2.14.0

# 資料庫
psycopg2-binary>=2.9.9  # PostgreSQL（生產環境）
# SQLite 為 Python 內建，無需安裝

# 排程器
APScheduler>=3.10.4

# Google Maps API（選用）
googlemaps>=4.10.0  # 官方 SDK（替代方案）
